refactor(test): move inline test modules to tests/ dirs
