+++ apps/gallery/src/patterns/document.rs
@@ -34,7 +34,7 @@ use gpui::{
 /// The document on the page. Canonical markdown — `serialize(parse(SOURCE))`
 /// returns it unchanged, which the gallery's tests assert, so the Source
 /// segment can be compared against this by eye.
-const SOURCE: &str = "\
+pub const SOURCE: &str = "\
 # Markdown
 
 Body text with **bold**, _italic_, ~~struck~~, `inline code`, and a \
@@ -210,20 +210,5 @@ impl Render for Document {
                             .child(div().max_w(px(680.0)).child(body)),
                     ),
             )
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// The Source segment shows `serialize(&doc)` and the page invites you to
-    /// compare it with [`SOURCE`]. If the constant drifts out of canonical form
-    /// the two stop matching, and the screen quietly demonstrates a round trip
-    /// it does not actually survive.
-    #[test]
-    fn the_source_is_canonical() {
-        let doc = bezel_markdown::parse(SOURCE);
-        assert_eq!(bezel_markdown::serialize(&doc), SOURCE);
     }
 }
